Refactor: 불필요한 폴더 삭제, 파일 이동
