+++ src/features/problemSet/utils/processMathHtml.ts
