+++ src/features/problemSet/utils/processMathText.ts
