+++ src/pages/ProblemListPage/components/ProblemItem.tsx
@@ -1,4 +1,4 @@
-import { formatBaekjoonTier } from '@/utils/converter/convertTier';
+import { formatBaekjoonTier } from '@/features/problemSet/utils/convertTier';
 import { useNavigate } from 'react-router-dom';
 
 interface IProblemItemProps {
